+++ packages/api/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @collectium/api
 
+## 0.2.3
+
+### Patch Changes
+
+- add builder package
+
+- Updated dependencies []:
+  - @collectium/core@0.2.3
+
 ## 0.2.2
 
 ### Patch Changes