+++ packages/lib/CHANGELOG.md
@@ -1,5 +1,16 @@
 # collectium
 
+## 0.2.3
+
+### Patch Changes
+
+- add builder package
+
+- Updated dependencies []:
+  - @collectium/check@0.2.3
+  - @collectium/core@0.2.3
+  - @collectium/cli@0.2.3
+
 ## 0.2.2
 
 ### Patch Changes